backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
psycopg2-binary
werkzeug
pydantic
requests
openai
PyJWT